BREAKING(text): align to single-export file pattern
