+++ text/mod.ts
@@ -23,4 +23,7 @@ export * from "./levenshtein_distance.ts";
 export * from "./closest_string.ts";
 export * from "./compare_similarity.ts";
 export * from "./word_similarity_sort.ts";
-export * from "./case.ts";
+export * from "./to_camel_case.ts";
+export * from "./to_kebab_case.ts";
+export * from "./to_pascal_case.ts";
+export * from "./to_snake_case.ts";
